Implement Build_Mode class and updated Feature_Flags class.
