+++ includes/Core/Util/Feature_Flags.php
@@ -21,16 +21,6 @@ use ArrayAccess;
  */
 class Feature_Flags {
 
-	const MODE_PRODUCTION = 'production';
-
-	/**
-	 * Feature flag mode.
-	 *
-	 * @since 1.22.0
-	 * @var string
-	 */
-	private static $mode = self::MODE_PRODUCTION;
-
 	/**
 	 * Feature flag definitions.
 	 *
@@ -52,12 +42,6 @@ class Feature_Flags {
 			return false;
 		}
 
-		$feature_modes = is_array( static::$features[ $feature ] ) ?
-			static::$features[ $feature ] :
-			array( static::$features[ $feature ] );
-
-		$feature_enabled = in_array( static::get_mode(), $feature_modes, true );
-
 		/**
 		 * Filters a feature flag's status (on or off).
 		 *
@@ -70,7 +54,7 @@ class Feature_Flags {
 		 * @param string $feature         The feature name.
 		 * @param string $mode            Site mode for loading features ('development' or 'production').
 		 */
-		return apply_filters( 'googlesitekit_is_feature_enabled', $feature_enabled, $feature, static::get_mode() );
+		return apply_filters( 'googlesitekit_is_feature_enabled', false, $feature );
 	}
 
 	/**
@@ -106,19 +90,6 @@ class Feature_Flags {
 	}
 
 	/**
-	 * Sets the feature flag mode.
-	 *
-	 * @since 1.22.0
-	 *
-	 * @param string $mode Feature flag mode.
-	 */
-	public static function set_mode( $mode ) {
-		if ( $mode && is_string( $mode ) ) {
-			static::$mode = $mode;
-		}
-	}
-
-	/**
 	 * Gets all available feature flags.
 	 *
 	 * @since 1.26.0
@@ -126,30 +97,7 @@ class Feature_Flags {
 	 * @return array An array of all available features.
 	 */
 	public static function get_available_features() {
-		$feature_keys = array();
-
-		if ( is_array( static::$features ) ) {
-			$feature_keys = array_keys( static::$features );
-		}
-
-		return $feature_keys;
+		return static::$features;
 	}
 
-	/**
-	 * Gets the current feature flag mode.
-	 *
-	 * @since 1.22.0
-	 *
-	 * @return string Current mode.
-	 */
-	private static function get_mode() {
-		/**
-		 * Filter the feature flag mode.
-		 *
-		 * @since 1.22.0
-		 *
-		 * @param string $mode The current feature flag mode.
-		 */
-		return (string) apply_filters( 'googlesitekit_flag_mode', static::$mode ) ?: self::MODE_PRODUCTION;
-	}
 }
