+++ includes/Plugin.php
@@ -10,6 +10,7 @@
 
 namespace Google\Site_Kit;
 
+use Google\Site_Kit\Core\Util\Build_Mode;
 use Google\Site_Kit\Core\Util\Feature_Flags;
 use Google\Site_Kit\Core\Util\JSON_File;
 
@@ -246,7 +247,7 @@ final class Plugin {
 		}
 
 		$config = new JSON_File( GOOGLESITEKIT_PLUGIN_DIR_PATH . 'dist/config.json' );
-		Feature_Flags::set_mode( $config['flagMode'] );
+		Build_Mode::set_mode( $config['buildMode'] );
 		Feature_Flags::set_features( (array) $config['features'] );
 
 		static::$instance = new static( $main_file );
@@ -254,4 +255,5 @@ final class Plugin {
 
 		return true;
 	}
+
 }
